Episode 8: Sprite Map
I had a hell of a time getting this to work. I originally created the
SpriteState struct as a class, so that gave me a KeyNotFound Exception
at runtime. It made sense once I went back and reviewed the differences
between a struct and a class...

Since a class is a reference type, when arrays (in this case a
dictionary) of that object are made, the elements of that array are just
references to the instances of that object, so that is why my dictionary
couldn't find the key.

A struct, on the other hand, is a value type. So when an array is made
the array elements are the actual values of the object (allocated
in-line) rather than just references.

Fascinating stuff